+++ Project2/PipelineSimulation.Core/Caches/SetEntry.cs
@@ -8,10 +8,10 @@ namespace PipelineSimulation.Core.Caches {
 	public class SetEntry {
 
 		public int Tag;
-		public int Data;
+		public ushort Data;
 		public int LRU;
 
-		public SetEntry(int tag, int data) {
+		public SetEntry(int tag, ushort data) {
 			Tag = tag;
 			Data = data;
 			LRU = 0;
